Add prestable sync workflow, move stable branches to config file (#21463)
