+++ .github/workflows/automerge_pr.yaml
@@ -6,6 +6,8 @@ on:
 concurrency:
   group: ${{ github.workflow }}
   cancel-in-progress: true
+env:
+  GH_TOKEN: ${{ secrets.YDBOT_TOKEN }}
 jobs:
   check-pr:
     runs-on: ubuntu-latest
@@ -16,8 +18,23 @@ jobs:
           sparse-checkout: |
             .github
             ydb/ci/
-      - uses: ./.github/actions/rightlib_sync
-        with:
-          command: check-pr
-          repository: ${{ github.repository }}
-          gh_personal_access_token: ${{ secrets.GH_PERSONAL_ACCESS_TOKEN }}
+      - name: install packages
+        shell: bash
+        run: |
+          pip install PyGithub==2.5.0
+
+      - name: configure
+        shell: bash
+        run: |
+            git config --global user.name YDBot
+            git config --global user.email ydbot@ydb.tech
+            git config --local github.token ${{ env.GH_TOKEN }}
+
+      - name: run-command
+        shell: bash
+        env:
+          REPO: ${{ github.repository }}
+          TOKEN: ${{ env.GH_TOKEN }}
+        run: |
+          cd ./ydb/ci/rightlib
+          ./automerge.py
