+++ .github/workflows/nightly_build.yml
@@ -18,10 +18,16 @@ jobs:
     outputs:
       branches: ${{ steps.set-branches.outputs.branches }}
     steps:
-      - id: set-branches
-        run: |
+    - name: Checkout
+      uses: actions/checkout@v4
+      with:
+        sparse-checkout: |
+          .github/config/stable_branches.json
+    - name: Read branches from config
+      id: set-branches
+      run: |
           if [[ "${{ github.event_name }}" == "schedule" || ("${{ inputs.use_default_branches }}" == "true"  && "${{ github.ref_type }}" != "tag") ]]; then
-            echo "branches=['main', 'stable-25-1', 'stable-24-4', 'stable-25-1-1', 'stable-25-1-analytics', 'stable-25-1-2', 'stable-25-1-3', 'prestable-25-2', 'prestable-25-3']" >> $GITHUB_OUTPUT
+            echo "branches=$(jq -c '.stables_for_nightly' .github/config/stable_branches.json)" >> $GITHUB_OUTPUT
           else
             echo "branches=['${{ github.ref_name }}']" >> $GITHUB_OUTPUT
           fi
