+++ .github/workflows/rightlib_sync.yml
@@ -6,18 +6,36 @@ on:
 concurrency:
   group: ${{ github.workflow }}
   cancel-in-progress: true
+env:
+  GH_TOKEN: ${{ secrets.YDBOT_TOKEN }}
 jobs:
   create-pr:
     runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        include:
+          - base_branch: prestable-25-2
+            head_branch: stable-25-1
+            label: sync-prestable
+          - base_branch: prestable-25-3
+            head_branch: main
+            label: sync-prestable
+          - base_branch: main
+            head_branch: rightlib
+            label: rightlib
     steps:
       - name: checkout
         uses: actions/checkout@v4
         with:
+          fetch-depth: 0
           sparse-checkout: |
             .github
             ydb/ci/
       - uses: ./.github/actions/rightlib_sync
         with:
-          command: create-pr
+          base_branch: ${{ matrix.base_branch }}
+          head_branch: ${{ matrix.head_branch }}
+          label: ${{ matrix.label }}
           repository: ${{ github.repository }}
-          gh_personal_access_token: ${{ secrets.GH_PERSONAL_ACCESS_TOKEN }}
+          gh_personal_access_token: ${{ env.GH_TOKEN }}
